Adding all the files in the project
The project itself is quite messy, with all sorts of files that might not even work anymore being added. The purpose is to add them and then remove
them later on while working on the project. It's about time we started version controlling this project. For now, we'll version control it so that
we can push it to a remote.